Fixed hl1 master server and hl2/rdfk masterserver, they should work as intended (again)
Started to move the servers from having the same udp and tcp code copied to each file to having the all the servers use a file called networkhandler.py. This way code is much cleaner and only packet handling code and the init for the server network handler is in each of the server files in steamemu.

updated more packet documentation to be more accurate